+++ src/components/BookingPanel/BookingPanel.example.js
@@ -4,16 +4,18 @@ import { LISTING_STATE_CLOSED } from '../../util/types';
 import BookingPanel from './BookingPanel';
 import css from './BookingPanelExample.css';
 
+const noop = () => null;
 export const Default = {
   component: BookingPanel,
   props: {
     className: css.example,
-    listing: createListing('listing_1'),
+    listing: createListing('listing_1', { availabilityPlan: { timezone: 'Etc/UTC' } }),
     onSubmit: values => console.log('Submit:', values),
     title: <span>Booking title</span>,
     subTitle: 'Hosted by Author N',
     authorDisplayName: 'Author Name',
     onManageDisableScrolling: () => null,
+    onFetchTimeSlots: noop,
   },
 };
 
@@ -21,11 +23,15 @@ export const WithClosedListing = {
   component: BookingPanel,
   props: {
     className: css.example,
-    listing: createListing('listing_1', { state: LISTING_STATE_CLOSED }),
+    listing: createListing('listing_1', {
+      availabilityPlan: { timezone: 'Etc/UTC' },
+      state: LISTING_STATE_CLOSED,
+    }),
     onSubmit: values => console.log('Submit:', values),
     title: <span>Booking title</span>,
     subTitle: 'Hosted by Author N',
     authorDisplayName: 'Author Name',
-    onManageDisableScrolling: () => null,
+    onManageDisableScrolling: noop,
+    onFetchTimeSlots: noop,
   },
 };