+++ src/components/EditListingWizard/EditListingWizard.js
@@ -264,6 +264,7 @@ class EditListingWizard extends Component {
                 handleCreateFlowTabScrolling={this.handleCreateFlowTabScrolling}
                 handlePublishListing={this.handlePublishListing}
                 fetchInProgress={fetchInProgress}
+                onManageDisableScrolling={onManageDisableScrolling}
               />
             );
           })}