+++ src/components/TransactionPanel/TransactionPanel.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import { array, arrayOf, bool, func, number, string } from 'prop-types';
+import { array, arrayOf, bool, func, number, object, string } from 'prop-types';
 import { FormattedMessage, injectIntl, intlShape } from '../../util/reactIntl';
 import classNames from 'classnames';
 import {
@@ -182,6 +182,7 @@ export class TransactionPanelComponent extends Component {
       sendMessageError,
       sendReviewInProgress,
       sendReviewError,
+      onFetchTimeSlots,
       onManageDisableScrolling,
       onShowMoreMessages,
       transactionRole,
@@ -193,8 +194,7 @@ export class TransactionPanelComponent extends Component {
       acceptSaleError,
       declineSaleError,
       onSubmitBookingRequest,
-      timeSlots,
-      fetchTimeSlotsError,
+      monthlyTimeSlots,
       nextTransitions,
     } = this.props;
 
@@ -457,8 +457,8 @@ export class TransactionPanelComponent extends Component {
                   authorDisplayName={authorDisplayName}
                   onSubmit={onSubmitBookingRequest}
                   onManageDisableScrolling={onManageDisableScrolling}
-                  timeSlots={timeSlots}
-                  fetchTimeSlotsError={fetchTimeSlotsError}
+                  monthlyTimeSlots={monthlyTimeSlots}
+                  onFetchTimeSlots={onFetchTimeSlots}
                 />
               ) : null}
               <BreakdownMaybe
@@ -500,8 +500,7 @@ TransactionPanelComponent.defaultProps = {
   savePaymentMethodFailed: false,
   sendMessageError: null,
   sendReviewError: null,
-  timeSlots: null,
-  fetchTimeSlotsError: null,
+  monthlyTimeSlots: null,
   nextTransitions: null,
 };
 
@@ -522,13 +521,13 @@ TransactionPanelComponent.propTypes = {
   sendMessageError: propTypes.error,
   sendReviewInProgress: bool.isRequired,
   sendReviewError: propTypes.error,
+  onFetchTimeSlots: func.isRequired,
   onManageDisableScrolling: func.isRequired,
   onShowMoreMessages: func.isRequired,
   onSendMessage: func.isRequired,
   onSendReview: func.isRequired,
   onSubmitBookingRequest: func.isRequired,
-  timeSlots: arrayOf(propTypes.timeSlot),
-  fetchTimeSlotsError: propTypes.error,
+  monthlyTimeSlots: object,
   nextTransitions: array,
 
   // Sale related props